working on the details screen

+++ AdidasProducts/AdidasProducts/Details/Model/ReviewModel.swift
@@ -10,6 +10,6 @@ import Foundation
 struct Review: Codable {
     var productId: String
     var locale: String
-    var rating: Double
+    var rating: Int
     var text: String
 }
